+++ DHL-Project2-Site/src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import { RouterModule, Routes } from '@angular/router'
+
 import { AppComponent } from './app.component';
-import { LoginComponent } from './login/login.component';
-import { CourseComponent } from './Course/course.component';
+import { LoginComponent } from '../login/login.component';
+import { CourseComponent } from '../course/course.component';
 
 
 const routes: Routes = [
