+++ DHL-Project2-Site/src/app/app.component.html
@@ -1,18 +1,20 @@
-<!-- Toolbar -->
+<!-- Toolbar-->
+
+
+
+
 <app-navbar></app-navbar>
-
 <div class="content" role="main">
-	<router-outlet></router-outlet>
-	<p *ngIf="user">Current User: {{user}}</p>
+  <router-outlet></router-outlet>
 
-	<h1>Welcome</h1>
-	<!-- Highlight Card -->
-	<div class="card highlight-card card-small">
-	  <span>{{ title }} app is running!</span>
-	</div>
+
+	<h1 *ngIf="user">Welcome {{user}}</h1>
+
+
+	<h1></h1>
 
   <button mat-button mat-button color="primary" routerLink="/login">Login</button>
   <button mat-button mat-button color="primary" routerLink="/class">Courses</button>
-  <!-- Footer -->
+   <!-- Footer -->
   <footer hidden=true></footer>
 </div>
