+++ DHL-Project2-Site/src/app/app.component.ts
@@ -6,7 +6,5 @@ import { Component } from '@angular/core';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  title = 'Course Portal';
-
   user = sessionStorage.getItem('currentEmail');
 }
