--- DHL-Project2-Site/src/app/navbar/navbar.component.html
@@ -1,7 +1,0 @@
-<nav>
-  <mat-toolbar class="toolbar" color="primary" role="banner">
-    <span class="panel-title" id='page-name' style="margin-left:10%">{{pageName}}</span>
-    <span class="spacer"></span>
-    <button mat-button id='login' style="margin-right:10%">Login</button>
-  </mat-toolbar>
-</nav>
