+++ DHL-Project2-Site/src/login/login.component.html
@@ -1,4 +1,4 @@
-<div class="login-page">
+  <div class="login-page">
     <div class="form">
       <form class="login-form" [formGroup]="myForm">
         <input type="email" placeholder="email" formControlName="email"/>
