+++ DHL-Project2-Site/src/navbar/navbar.component.ts
@@ -1,12 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 
+import { User } from '../shared/Models/user.model';
+
 @Component({
   selector: 'app-navbar',
   templateUrl: './navbar.component.html',
   styleUrls: ['./navbar.component.css']
 })
 export class NavbarComponent implements OnInit {
-  pageName = 'Welcome';
+  user: User;
 
   constructor() { }
