+++ DHL-Project2-Site/src/shared/Models/user.model.ts
